+++ settings.gradle.kts
@@ -31,3 +31,7 @@ include("tusow-service")
 include("tusow-cli")
 include("tusow-grpc")
 include("tusow-full")
+include("tusow-grpc-presentation")
+include("tusow-grpc-text")
+include("tusow-grpc-text")
+include("tusow-grpc-logic")
