+++ tusow-grpc/build.gradle.kts
@@ -38,6 +38,9 @@ dependencies {
     api(project(":linda-text"))
     api(project(":linda-text-presentation"))
     api(project(":tusow-service"))
+    api(project(":tusow-grpc-presentation"))
+    api(project(":tusow-grpc-text"))
+    api(project(":tusow-grpc-logic"))
     api("io.vertx:vertx-core:4.2.2")
     api("io.vertx:vertx-web:4.2.2")
     api("io.grpc:grpc-protobuf:${grpcVersion}")
